Fix review comments, Iter. 2 -- MetadataChainExt: absorb accept* API
